Move the interpreters module to the official stdlib.
