+++ Lib/test/test_threading.py
@@ -28,7 +28,7 @@ from test import lock_tests
 from test import support
 
 try:
-    from test.support import interpreters
+    import interpreters
 except ImportError:
     interpreters = None
