+++ Makefile.pre.in
@@ -2526,6 +2526,7 @@ LIBSUBDIRS=	asyncio \
 		http \
 		idlelib idlelib/Icons \
 		importlib importlib/resources importlib/metadata \
+		interpreters \
 		json \
 		logging \
 		multiprocessing multiprocessing/dummy \
@@ -2573,7 +2574,6 @@ TESTSUBDIRS=	idlelib/idle_test \
 		test/subprocessdata \
 		test/support \
 		test/support/_hypothesis_stubs \
-		test/support/interpreters \
 		test/test_asyncio \
 		test/test_capi \
 		test/test_cext \
